Updated sort, filter events.
